recipe_rag_backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.0.346
openai==0.28.1
faiss-cpu==1.7.4
pandas==2.1.3
python-dotenv==1.0.0
pydantic==2.5.0
tiktoken==0.9.0
